remove the -o option; update docs and tests

+++ tool/grind.dart
@@ -66,17 +66,17 @@ void updateGhPages(GrinderContext context) {
 void testGenerators(GrinderContext context) {
   Directory fooDir = new Directory('foo');
 
-  if (fooDir.existsSync()) {
-    fooDir.deleteSync(recursive: true);
-  }
-
   try {
     for (stagehand.Generator generator in stagehand.generators) {
+      if (fooDir.existsSync()) fooDir.deleteSync(recursive: true);
+      fooDir.createSync();
+
       context.log('');
       context.log('${generator.id} template:');
 
-      runDartScript(context, 'bin/stagehand.dart',
-          arguments: ['--mock-analytics', '-o', 'foo', generator.id]);
+      runDartScript(context, '../bin/stagehand.dart',
+          arguments: ['--mock-analytics', generator.id],
+          workingDirectory: fooDir.path);
 
       File file = joinFile(fooDir, [generator.entrypoint.path]);
